restructure to fit package
