+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
   - 'packages/*'
-  - 'apps/*'
+  - 'samples/*'
+  - 'samples/widgets/*'
