+++ samples/react/.browserslistrc
