+++ samples/react/package.json
