+++ samples/react/public/favicon.png
