+++ samples/react/public/index.html
