+++ samples/react/public/index.js
