+++ samples/react/src/Layout.tsx
@@ -1,7 +1,7 @@
 import { Button, Div, Flex } from "@workleap/orbiter-ui";
 import { Link, Outlet } from "react-router-dom";
 
-import { MovieFinder, SelectedMovie } from "@workleap/movie-widgets/react";
+import { MovieFinder, SelectedMovie } from "../../widgets/movie-widgets/dist/react/index.js";
 
 import { AppContextProvider, useAppContext } from "./AppContext.tsx";
