+++ samples/react/swc.build.js
