+++ samples/react/swc.dev.js
