+++ samples/react/webpack.build.js
