+++ samples/react/webpack.dev.js
