+++ samples/vanilla-js/.eslintignore
