+++ samples/vanilla-js/.eslintrc.json
