+++ samples/vanilla-js/package.json
