+++ samples/vanilla-js/public/index.html
