+++ samples/vanilla-js/public/index.js
