+++ samples/vanilla-js/server.js
