+++ samples/widgets/movie-widgets/.eslintignore
