+++ samples/widgets/movie-widgets/.eslintrc.json
