+++ samples/widgets/movie-widgets/package.json
