+++ samples/widgets/movie-widgets/src/AppContextProvider.tsx
