+++ samples/widgets/movie-widgets/src/MovieDetails.tsx
@@ -13,7 +13,7 @@ import {
     ThemeProvider
 } from "@workleap/orbiter-ui";
 import { useEffect, useState } from "react";
-import { type MovieData, useAppContext } from "./AppContextProvider.tsx";
+import { type MovieData, useAppContext } from "./AppContextProvider.js";
 
 export interface MovieDetailsProps {
     showRanking: boolean;
