+++ samples/widgets/movie-widgets/src/MovieFinderElement.tsx
@@ -1,5 +1,5 @@
 import { WebComponentHTMLElement } from "@workleap/r2wc";
-import { MovieFinder } from "./MovieFinder.tsx";
+import { MovieFinder } from "./MovieFinder.js";
 
 export class MovieFinderElement extends WebComponentHTMLElement {
     get reactComponent() {
