+++ samples/widgets/movie-widgets/src/SelectedMovie.tsx
@@ -1,6 +1,6 @@
 import { VideoCameraIcon } from "@hopper-ui/icons";
 import { Div, ThemeProvider } from "@workleap/orbiter-ui";
-import { useAppContext } from "./AppContextProvider.tsx";
+import { useAppContext } from "./AppContextProvider.js";
 
 export function SelectedMovie() {
     const { selectedMovie, theme } = useAppContext();
