+++ samples/widgets/movie-widgets/src/helpers/react/types.d.ts
