+++ samples/widgets/movie-widgets/src/index.css
