+++ samples/widgets/movie-widgets/src/index.ts
