+++ samples/widgets/movie-widgets/src/types.d.ts
@@ -1,5 +1,5 @@
 import type { IWidgetsManager } from "@workleap/r2wc";
-import type { AppSettings } from "./AppContextProvider.tsx";
+import type { AppSettings } from "./AppContextProvider.js";
 
 export declare global {
     interface Window {
