+++ samples/widgets/movie-widgets/tsconfig.json
