+++ samples/widgets/movie-widgets/tsup.build.cdn.ts
