+++ samples/widgets/movie-widgets/tsup.build.react.ts
